pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.backends.legacy:build"

[project]
name = "zipjeweler"
version = "0.1.0"
description = "AI-powered marketing team for ZipJeweler — 36 autonomous agents across 7 crews"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.11"
dependencies = [
    "crewai>=0.28.0",
    "crewai-tools>=0.1.0",
    "anthropic>=0.25.0",
    "streamlit>=1.32.0",
    "gspread>=6.0.0",
    "google-auth>=2.28.0",
    "python-dotenv>=1.0.0",
    "typer>=0.12.0",
    "rich>=13.7.0",
    "praw>=7.7.0",
    "tweepy>=4.14.0",
    "instagrapi>=2.0.0",
    "requests>=2.31.0",
    "apscheduler>=3.10.0",
    "pandas>=2.2.0",
    "httpx>=0.27.0",
    "pillow>=10.3.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "black>=24.0.0",
    "ruff>=0.4.0",
]

[project.scripts]
zipjeweler = "zipjeweler.cli:app"

[tool.setuptools.packages.find]
where = ["."]
include = ["zipjeweler*"]
